one_last_time/requirements.txt
Flask
Flask-Login
Flask-SQLAlchemy
scikit-learn
pandas
nltk
spacy
pdfminer.six
python-docx


